requirements.txt
Flask==3.0.0
openai==0.28.1
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
pytest==7.4.3
pytest-flask==1.3.0
